Update courier dependency to 0.6.0 (#3)
This update requires removing references to the EmptyContent type as
well as ensuring the courier can deliver a templated email with a null
subject. Because MailJet does not support using template variables in
the subject, the courier will send an empty string when it is null to
avoid validation errors, and the subject on the template will still be
used.